requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
python-multipart
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
pypdf
qdrant-client
sentence-transformers
# FastAPI & Server
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.3

# LangChain
langchain==0.1.20
langchain-groq==0.1.3
langchain-huggingface==0.0.1
langchain-community==0.0.38
langchain-core==0.1.52

# Embeddings & Vector DB
sentence-transformers==2.5.1
qdrant-client==1.7.3

# Document Processing
pypdf==4.0.1

# AI/ML
transformers==4.38.2
torch==2.2.1

# Utilities
requests==2.31.0